% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_inv_and_port_names_if_needed}
\alias{add_inv_and_port_names_if_needed}
\title{Add investor and portfolio names to a data frame if needed}
\usage{
add_inv_and_port_names_if_needed(data, portfolio_name, investor_name)
}
\arguments{
\item{data}{Data frame to add the names to.}

\item{portfolio_name}{Portfolio name value.}

\item{investor_name}{Investor name value.}
}
\value{
Data frame with the names added if needed.
}
\description{
Add investor and portfolio names to a data frame if needed
}
