% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_executive_summary.R
\name{prepare_executive_summary}
\alias{prepare_executive_summary}
\title{Prepare an Executive Summary PDF}
\usage{
prepare_executive_summary(
  summary_output_dir,
  survey_dir,
  real_estate_dir,
  score_card_dir,
  project_code,
  language_select,
  peer_group,
  investor_name,
  portfolio_name,
  start_year,
  currency_exchange_value,
  total_portfolio,
  equity_results_portfolio,
  bonds_results_portfolio,
  indices_equity_results_portfolio,
  indices_bonds_results_portfolio,
  audit_file,
  emissions,
  peers_bonds_results_portfolio,
  peers_bonds_results_user,
  peers_equity_results_portfolio,
  peers_equity_results_user
)
}
\arguments{
\item{summary_output_dir}{Directory where the executive summary will be
saved.}

\item{survey_dir}{Directory containing survey data.}

\item{real_estate_dir}{Directory containing real estate data.}

\item{score_card_dir}{Directory containing score card data.}

\item{project_code}{Project code, defining the template to use.}

\item{language_select}{Language to use for the executive summary.}

\item{peer_group}{Investor's Peer group.}

\item{investor_name}{Investor's name.}

\item{portfolio_name}{Portfolio's name.}

\item{start_year}{Start year of the analysis.}

\item{currency_exchange_value}{Exchange value from Display cyurrency to USD.}

\item{total_portfolio}{Data frame with the total portfolio information.}

\item{equity_results_portfolio}{Data frame with the equity results for the
portfolio.}

\item{bonds_results_portfolio}{Data frame with the bonds results for the
portfolio.}

\item{indices_equity_results_portfolio}{Data frame with the equity results
for the indices (benchmarks).}

\item{indices_bonds_results_portfolio}{Data frame with the bonds results for
the indices (benchmarks).}

\item{audit_file}{Audit file.}

\item{emissions}{Emissions data.}

\item{peers_bonds_results_portfolio}{Data frame with the portfolio-level
bonds results for the peers.}

\item{peers_bonds_results_user}{Data frame with the user-level bonds results
for the peers.}

\item{peers_equity_results_portfolio}{Data frame with the portfolio-level
equity results for the peers.}

\item{peers_equity_results_user}{Data frame with the user-level equity
results for the peers.}
}
\value{
List of manifest information, suitable for use in
\code{pacta.workflow.utils::export_manifest}.
}
\description{
Run the required steps to prepare an executive summary from existing
analysis results.
}
