% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_interactive_report.R
\name{prepare_interactive_report}
\alias{prepare_interactive_report}
\title{Prepare an Interactive HTML Report}
\usage{
prepare_interactive_report(
  report_output_dir,
  project_report_name,
  survey_dir,
  real_estate_dir,
  language_select,
  peer_group,
  investor_name,
  portfolio_name,
  start_year,
  currency_exchange_value,
  select_scenario,
  scenario_other,
  portfolio_allocation_method,
  scenario_geography,
  sector_list,
  green_techs,
  tech_roadmap_sectors,
  pacta_sectors_not_analysed,
  display_currency,
  audit_file,
  emissions,
  portfolio_overview,
  equity_results_portfolio,
  bonds_results_portfolio,
  equity_results_company,
  bonds_results_company,
  equity_results_map,
  bonds_results_map,
  indices_bonds_results_portfolio,
  indices_equity_results_portfolio,
  peers_bonds_results_portfolio,
  peers_bonds_results_user,
  peers_equity_results_portfolio,
  peers_equity_results_user,
  analysis_output_manifest
)
}
\arguments{
\item{report_output_dir}{Directory where the interactive report will be
saved.}

\item{project_report_name}{Name of the project.}

\item{survey_dir}{Directory containing survey data.}

\item{real_estate_dir}{Directory containing real estate data.}

\item{language_select}{Language to use for the interactive report.}

\item{peer_group}{Investor's Peer group.}

\item{investor_name}{Investor's name.}

\item{portfolio_name}{Portfolio's name.}

\item{start_year}{Start year of the analysis.}

\item{currency_exchange_value}{Exchange value from Display currency to USD.}

\item{select_scenario}{Default scenario to be displayed in visualizations.}

\item{scenario_other}{Other scenarios to be displayed in visualizations.}

\item{portfolio_allocation_method}{Method used to allocate the portfolio.}

\item{scenario_geography}{Geographies included in the scenarios.}

\item{sector_list}{Sectors included in the analysis.}

\item{green_techs}{Green technologies included in the analysis.}

\item{tech_roadmap_sectors}{Sectors included in the technology roadmap.}

\item{pacta_sectors_not_analysed}{Sectors not analysed in the analysis.}

\item{display_currency}{Currency to be displayed in the report.}

\item{audit_file}{Audit file.}

\item{emissions}{Emissions data.}

\item{portfolio_overview}{Portfolio overview data.}

\item{equity_results_portfolio}{Data frame with the equity results for the
portfolio.}

\item{bonds_results_portfolio}{Data frame with the bonds results for the
portfolio.}

\item{equity_results_company}{Data frame with company-level equity results
for the portfolio}

\item{bonds_results_company}{Data frame with company-level bonds results for
the portfolio}

\item{equity_results_map}{Data frame with the equity results for the map.}

\item{bonds_results_map}{Data frame with the bonds results for the map.}

\item{indices_bonds_results_portfolio}{Data frame with the bonds results for
the indices (benchmarks).}

\item{indices_equity_results_portfolio}{Data frame with the equity results
for the indices (benchmarks).}

\item{peers_bonds_results_portfolio}{Data frame with the portfolio-level
bonds results for the peers.}

\item{peers_bonds_results_user}{Data frame with the user-level bonds results
for the peers.}

\item{peers_equity_results_portfolio}{Data frame with the portfolio-level
equity results for the peers.}

\item{peers_equity_results_user}{Data frame with the user-level equity
results for the peers.}

\item{analysis_output_manifest}{List of manifest information from the
analysis.}
}
\value{
List of manifest information, suitable for use in
\code{pacta.workflow.utils::export_manifest}.
}
\description{
Run the required steps to prepare an interactive HTML report from existing
analysis results.
}
