% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_rds_or_return_alt_data}
\alias{read_rds_or_return_alt_data}
\title{Read an RDS file or return an alternative data set}
\usage{
read_rds_or_return_alt_data(filepath, alt_return = NULL)
}
\arguments{
\item{filepath}{Path to the RDS file to read.}

\item{alt_return}{Data to return if the file is not found.}
}
\value{
Data read from the RDS file or the alternative data set.
}
\description{
Read an RDS file or return an alternative data set
}
