% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pacta_reporting_process.R
\name{run_pacta_reporting_process}
\alias{run_pacta_reporting_process}
\title{Run PACTA Report and Summary generation process}
\usage{
run_pacta_reporting_process(
  params,
  analysis_output_dir = Sys.getenv("ANALYSIS_OUTPUT_DIR"),
  benchmarks_dir = Sys.getenv("BENCHMARKS_DIR"),
  report_output_dir = Sys.getenv("REPORT_OUTPUT_DIR"),
  summary_output_dir = Sys.getenv("SUMMARY_OUTPUT_DIR"),
  real_estate_dir = Sys.getenv("REAL_ESTATE_DIR"),
  survey_dir = Sys.getenv("SURVEY_DIR"),
  score_card_dir = Sys.getenv("SCORE_CARD_DIR")
)
}
\arguments{
\item{params}{List of parameters for the PACTA reporting. See JSON schema
for more details.}

\item{analysis_output_dir}{Directory containing the PACTA analysis results.}

\item{benchmarks_dir}{Directory containing the benchmark analysis results.}

\item{report_output_dir}{Directory where the interactive report will be
saved.}

\item{summary_output_dir}{Directory where the executive summary will be
saved.}

\item{real_estate_dir}{Directory containing real estate data.}

\item{survey_dir}{Directory containing survey data.}

\item{score_card_dir}{Directory containing score card data.}
}
\value{
List of manifest information, suitable for use in
\code{pacta.workflow.utils::export_manifest}.
}
\description{
Run the required steps to generate a PACTA report and executive summary.
}
