% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.pacta.report-package.R
\docType{package}
\name{workflow.pacta.report-package}
\alias{workflow.pacta.report}
\alias{workflow.pacta.report-package}
\title{workflow.pacta.report: Reporting functionality for PACTA}
\description{
Create Interactive Reports and Executive Summaries for PACTA
}
\author{
\strong{Maintainer}: Alex Axthelm \email{aaxthelm@rmi.org} (\href{https://orcid.org/0000-0001-8579-8565}{ORCID}) [contractor]

Authors:
\itemize{
  \item Jackson Hoffart \email{jackson.hoffart@gmail.com} (\href{https://orcid.org/0000-0002-8600-5042}{ORCID}) [contractor]
}

Other contributors:
\itemize{
  \item RMI \email{PACTA4investors@rmi.org} [copyright holder, funder]
}

}
\keyword{internal}
